classloader task
After taskdef supports the loaderref attribute, I've written a task,
that is able to explicitely create classloaders, append classpath
entries to existing classloaders and put the actual path of a
classloader into a property.

In several projects, this avoids some of the annoying constraints and
workarounds occasionally reported in bugzilla and the mailing lists.

I've merged it with Costin's "sleeping" classloader task and I think
it could be a helpful enhancement.

Currently it supports URLClassLoader and AntClassLoader and it is
designed to simply support custom extensions for any arbitrary
ClassLoader.

http://article.gmane.org/gmane.comp.jakarta.ant.devel/28676